% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{The Server Function}
\usage{
server(input, output)
}
\arguments{
\item{input}{standard shiny parameter}

\item{output}{standard shiny parameter}
}
\value{
None
}
\description{
The Server Function
}
