% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_methods.R
\name{seuratClustering}
\alias{seuratClustering}
\title{Seurat (SNN Graph Based) Clustering Function}
\usage{
seuratClustering(sO, noDim, res, reductionMethod, method)
}
\arguments{
\item{sO}{the underlying Seurat object}

\item{noDim}{the number of dimensions to be use in the clustering step}

\item{res}{the Seurat resolution parameter}

\item{reductionMethod}{the type of graph (UMAP or T-SNE) used to represent the results}

\item{method}{the technique used to compress the data (PCA or GLM-PCA)}
}
\description{
Seurat (SNN Graph Based) Clustering Function
}
